Move constants, spot attributes, and validated tables into new module
We're starting to see some circular dependency issues, and that generally means we need to organize our code better.  Imports need to be disentangled, and one way to do that is to structure the code so that the imports go in one direction.  I've created a new module, called types, which encompasses constants, spot attributes, and validated tables.  These modules have no external dependencies that are within the starfish code base, which is what we need!
