+++ examples/build_sample_experiment.py
@@ -2,7 +2,7 @@ import argparse
 import json
 
 from examples.support import AUX_IMAGE_NAMES, write_experiment_json
-from starfish.constants import Indices
+from starfish.types import Indices
 from starfish.util.argparse import FsExistsType
