+++ starfish/image/_registration/fourier_shift.py
@@ -4,8 +4,8 @@ import numpy as np
 from scipy.ndimage import fourier_shift
 from skimage.feature import register_translation
 
-from starfish.constants import Indices
 from starfish.stack import ImageStack
+from starfish.types import Indices
 from starfish.util.argparse import FsExistsType
 from ._base import RegistrationAlgorithmBase
