+++ starfish/image/_segmentation/watershed.py
@@ -8,11 +8,11 @@ from showit import image
 from skimage.feature import peak_local_max
 from skimage.morphology import watershed
 
-from starfish.constants import Indices
-from starfish.stack import ImageStack
-from starfish.munge import relabel
 from starfish.image._filter.util import bin_thresh, bin_open
+from starfish.munge import relabel
+from starfish.stack import ImageStack
 from starfish.stats import label_to_regions
+from starfish.types import Indices
 from ._base import SegmentationAlgorithmBase
