+++ starfish/spots/__init__.py
@@ -1,4 +1,3 @@
-from ._spot_attributes import SpotAttributes
 from ._detector import SpotFinder
 from ._decoder import Decoder
 from ._target_assignment import TargetAssignment
