+++ starfish/spots/_detector/__init__.py
@@ -1,10 +1,10 @@
 import argparse
 import os
 
-from starfish.stack import ImageStack
 from starfish.pipeline.pipelinecomponent import PipelineComponent
+from starfish.stack import ImageStack
 from starfish.util.argparse import FsExistsType
-from starfish.constants import Indices
+from starfish.types import Indices
 from . import _base
 from . import gaussian
 from . import local_max_peak_finder
