+++ starfish/spots/_detector/combine_adjacent_features.py
@@ -4,10 +4,9 @@ import numpy as np
 import pandas as pd
 from skimage.measure import regionprops, label
 
-from starfish.constants import Indices, Features
 from starfish.intensity_table import IntensityTable
 from starfish.munge import dataframe_to_multiindex
-from starfish.types import Number
+from starfish.types import Features, Indices, Number
 
 
 class ConnectedComponentDecodingResult(NamedTuple):
