+++ starfish/spots/_detector/gaussian.py
@@ -5,11 +5,9 @@ import pandas as pd
 import xarray as xr
 from skimage.feature import blob_log
 
-from starfish.constants import Features
 from starfish.intensity_table import IntensityTable
-from starfish.spots import SpotAttributes
 from starfish.stack import ImageStack
-from starfish.types import Number
+from starfish.types import Features, Number, SpotAttributes
 from starfish.util.argparse import FsExistsType
 from ._base import SpotFinderAlgorithmBase
 from .detect import measure_spot_intensity, detect_spots
