+++ starfish/spots/_detector/local_max_peak_finder.py
@@ -5,8 +5,8 @@ import xarray as xr
 from trackpy import locate
 
 from starfish.intensity_table import IntensityTable
-from starfish.spots import SpotAttributes
 from starfish.stack import ImageStack
+from starfish.types import SpotAttributes
 from ._base import SpotFinderAlgorithmBase
 from .detect import detect_spots
