+++ starfish/test/full_pipelines/api/test_iss_api.py
@@ -3,8 +3,8 @@ from starfish.image._filter.white_tophat import WhiteTophat
 from starfish.image._registration.fourier_shift import FourierShiftRegistration
 from starfish.util.synthesize import SyntheticData
 import numpy as np
-from starfish.constants import Indices
 from starfish.stack import ImageStack
+from starfish.types import Indices
 
 
 def test_iss_pipeline():
