+++ starfish/test/full_pipelines/api/test_merfish.py
@@ -1,10 +1,10 @@
 import pytest
 
-from starfish.constants import Indices
 from starfish.spots._detector.pixel_spot_detector import PixelSpotDetector
 from starfish.image._filter.gaussian_high_pass import GaussianHighPass
 from starfish.image._filter.gaussian_low_pass import GaussianLowPass
 from starfish.image._filter.richardson_lucy_deconvolution import DeconvolvePSF
+from starfish.types import Indices
 
 
 @pytest.mark.skip("TODO ambrosejcarr: fix this test in a future PR. Currently no spots are "
