+++ starfish/test/image/test_imagestack.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pytest
 
-from starfish.constants import Indices, Features
-from starfish.stack import ImageStack
 from starfish.intensity_table import IntensityTable
+from starfish.stack import ImageStack
+from starfish.types import Indices, Features
 # don't inspect pytest fixtures in pycharm
 # noinspection PyUnresolvedReferences
 from starfish.test.dataset_fixtures import (
