+++ starfish/test/pipeline/test_intensity_table.py
@@ -3,11 +3,11 @@ from typing import Tuple
 import numpy as np
 import pandas as pd
 
-from starfish.constants import Indices, Features
-from starfish.intensity_table import IntensityTable
 from starfish.codebook import Codebook
-from starfish.stack import ImageStack
+from starfish.intensity_table import IntensityTable
 from starfish.spots._detector.combine_adjacent_features import combine_adjacent_features
+from starfish.stack import ImageStack
+from starfish.types import Indices, Features
 # don't inspect pytest fixtures in pycharm
 # noinspection PyUnresolvedReferences
 from starfish.test.dataset_fixtures import (
