+++ starfish/types/__init__.py
@@ -1,0 +1,6 @@
+from typing import Union
+
+from ._constants import Coordinates, Features, Indices
+from ._spot_attributes import SpotAttributes
+
+Number = Union[int, float]
