+++ starfish/types/_constants.py
