+++ starfish/types/_spot_attributes.py
